Social_Topic_Insight_python/requirements.txt
# --- Web Framework & API ---
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart

# --- Database & Config ---
pymongo>=4.6.0
pika>=1.3.2
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# --- Crawler & Network ---
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=5.1.0
DrissionPage>=4.0.0
urllib3

# --- NLP & AI Core ---
# 基础 NLP
jieba>=0.42.1

# 深度学习与向量化 (Torch通常由sentence-transformers自动安装，但建议显式列出)
torch
transformers>=4.36.0
sentence-transformers>=2.2.2

# 聚类与主题建模 (BERTopic 核心依赖)
bertopic>=0.16.0
scikit-learn>=1.3.0
umap-learn>=0.5.5
hdbscan>=0.8.33
pandas>=2.1.0
numpy>=1.26.0
